refactor(mailer-lite): move auth out of the piece entry to break a circular import
Every action, the common props module and the trigger factory imported mailerLiteAuth
from '../..', while src/index.ts imported all of them back. The cycle only survives
because index.ts is the normal entry point: importing any action directly fails with
"Cannot read properties of undefined (reading 'name')" inside createPiece, because the
action bindings are still undefined when the piece is constructed.

Auth now lives in src/lib/auth.ts, the pattern every other piece follows. index.ts
re-exports mailerLiteAuth so nothing downstream changes.
