+++ packages/pieces/community/mailer-lite/src/lib/actions/find-subscriber.ts
@@ -1,6 +1,7 @@
 import MailerLite from '@mailerlite/mailerlite-nodejs';
 import { Property, createAction } from '@activepieces/pieces-framework';
-import { mailerLiteAuth } from '../..';
+import { mailerLiteAuth } from '../auth';
+import { findSubscriberOutputSchema } from '../output-schemas';
 
 export const findSubscriberAction = createAction({
 	auth: mailerLiteAuth,
@@ -14,6 +15,7 @@ export const findSubscriberAction = createAction({
 			'Look up a single MailerLite subscriber by their email address or subscriber ID. Use this to resolve a contact to their record (status, fields, groups) before updating or segmenting them. Read-only and idempotent.',
 		idempotent: true,
 	},
+	outputSchema: findSubscriberOutputSchema,
 	props: {
 		searchValue: Property.ShortText({
 			displayName: 'Subscriber ID or Email',
